pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["onthemod <onthemod@yandex.ru>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.2.3"
gunicorn = "^20.1.0"
postgres = "^4.0"
psycopg2-binary = "^2.9.6"
validators = "^0.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
